Moved Dashboard fragments to their own directory.
