+++ app/src/main/java/edu/uco/schambers4/octane/Activities/MainActivity.java
@@ -18,10 +18,10 @@ import android.widget.FrameLayout;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
-import edu.uco.schambers4.octane.Fragments.Workout.ExerciseListFragment;
-import edu.uco.schambers4.octane.Fragments.DashboardFragment;
+import edu.uco.schambers4.octane.Fragments.Dashboard.DashboardFragment;
 import edu.uco.schambers4.octane.Fragments.Ingredients.IngredientsFragment;
 import edu.uco.schambers4.octane.Fragments.Recipes.RecipeListFragment;
+import edu.uco.schambers4.octane.Fragments.Workout.ExerciseListFragment;
 import edu.uco.schambers4.octane.Fragments.Workout.WorkoutListFragment;
 import edu.uco.schambers4.octane.Models.Workout.ExerciseContainer;
 import edu.uco.schambers4.octane.R;
