+++ app/src/main/res/layout/fragment_details_for_date.xml
@@ -2,7 +2,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.uco.schambers4.octane.Fragments.DetailsForDateFragment"
+    tools:context="edu.uco.schambers4.octane.Fragments.Dashboard.DetailsForDateFragment"
     android:background="@drawable/layout_border"
     android:layout_marginTop="10dp">
